1. 优化用户登录信息

+++ 03-ww-auth-3100/src/main/java/cn/wickson/tech/collective/auth/service/AbstractAuthService.java
@@ -1,6 +1,7 @@
 package cn.wickson.tech.collective.auth.service;
 
 import cn.hutool.core.util.ObjUtil;
+import cn.hutool.core.util.StrUtil;
 import cn.wickson.tech.collective.auth.enums.ResultCodeAuth;
 import cn.wickson.tech.collective.common.constant.CacheConstants;
 import cn.wickson.tech.collective.common.enums.UseStatusEnum;
@@ -8,6 +9,8 @@ import cn.wickson.tech.collective.common.exception.UserOperationException;
 import cn.wickson.tech.collective.common.redis.service.RedisService;
 import cn.wickson.tech.collective.system.api.SystemUserApi;
 import cn.wickson.tech.collective.system.dto.AdminUserDTO;
+import cn.wickson.tech.collective.system.enums.LoginLogTypeEnum;
+import cn.wickson.tech.collective.system.enums.LoginResultEnum;
 import org.springframework.beans.factory.annotation.Value;
 
 import javax.annotation.Resource;
@@ -38,18 +41,27 @@ public class AbstractAuthService {
      * @param code             uuid
      * @param verificationCode 验证码
      */
-    protected void validCaptcha(String code, String verificationCode) {
+    protected void validCaptcha(String username, String code, String verificationCode) {
         // 如果没有开启验证码则不进行校验
         if (!Boolean.TRUE.equals(enable)) {
             return;
         }
 
-        // 验证验证码是否正确
         String captchaCodeKey = getCaptchaCodeKey(code);
         String verifyCode = redisService.getCacheObject(captchaCodeKey);
+        // 验证验证码是否存在
+        if (StrUtil.isBlankIfStr(verifyCode)) {
+            // 记录日志
+            loginLogService.createLoginLog(null, username, LoginLogTypeEnum.LOGIN_USERNAME, LoginResultEnum.CAPTCHA_NOT_FOUND);
+            throw UserOperationException.getInstance(ResultCodeAuth.CAPTCHA_CODE_ERROR);
+        }
+
+        // 验证验证码是否正确
         if (!verificationCode.equals(verifyCode)) {
             // 验证错误，删除验证码
             redisService.deleteObject(captchaCodeKey);
+            // 记录日志
+            loginLogService.createLoginLog(null, username, LoginLogTypeEnum.LOGIN_USERNAME, LoginResultEnum.CAPTCHA_CODE_ERROR);
             throw UserOperationException.getInstance(ResultCodeAuth.CAPTCHA_CODE_ERROR);
         }
 
@@ -64,14 +76,20 @@ public class AbstractAuthService {
     protected void validUserInfo(AdminUserDTO userInfo, String password) {
         // 当前用户是否存在
         if (ObjUtil.isNull(userInfo)) {
+            // 记录日志
+            loginLogService.createLoginLog(null, userInfo.getUsername(), LoginLogTypeEnum.LOGIN_USERNAME, LoginResultEnum.USER_NOT_FOUND);
             throw UserOperationException.getInstance(ResultCodeAuth.USER_NOT_FOUND);
         }
         // 当前用户密码是否正确
         if (!password.equals(userInfo.getPassword())) {
+            // 记录日志
+            loginLogService.createLoginLog(null, userInfo.getUsername(), LoginLogTypeEnum.LOGIN_USERNAME, LoginResultEnum.BAD_CREDENTIALS);
             throw UserOperationException.getInstance(ResultCodeAuth.USER_PASSWORD_ERROR);
         }
         // 是否被禁用
         if (UseStatusEnum.DISABLE.getStatus().equals(userInfo.getStatus())) {
+            // 记录日志
+            loginLogService.createLoginLog(null, userInfo.getUsername(), LoginLogTypeEnum.LOGIN_USERNAME, LoginResultEnum.USER_DISABLED);
             throw UserOperationException.getInstance(ResultCodeAuth.USER_STATUS_DISABLE);
         }
     }
